core: split Darwin warning fixes
Restore darwin_usb.c to this branch's merge-base state so #1904 remains focused on the promoted-enum va_start issue in core.c.

The Apple request-type casts are now isolated in #1914, while the language-ID conversion was already merged through #1903.

Related to #1903.

Related to #1904.

Related to #1914.

+++ libusb/os/darwin_usb.c
@@ -1130,8 +1130,7 @@ static int darwin_fetch_string_descriptor (struct libusb_context *ctx,
   langid = dpriv->langid;
   if (0 == langid) {
     memset (desc, 0, sizeof(desc));
-    req.bmRequestType = USBmakebmRequestType ((UInt8)kUSBIn,
-      (UInt8)kUSBStandard, (UInt8)kUSBDevice);
+    req.bmRequestType = USBmakebmRequestType (kUSBIn, kUSBStandard, kUSBDevice);
     req.bRequest      = kUSBRqGetDescriptor;
     req.wValue        = (UInt16)(LIBUSB_DT_STRING << 8);
     req.wIndex        = 0;
@@ -1141,15 +1140,14 @@ static int darwin_fetch_string_descriptor (struct libusb_context *ctx,
     req.completionTimeout = 100;
     kresult = (*device)->DeviceRequestTO (device, &req);
     if (kIOReturnSuccess == kresult && req.wLenDone >= 4) {
-      langid = (UInt16)((UInt16)desc[2] | ((UInt16)desc[3] << 8));
+      langid = (UInt16)desc[2] | ((UInt16)desc[3] << 8);
       dpriv->langid = langid;
     }
   }
 
   /* Fetch the requested string descriptor. */
   memset (desc, 0, sizeof(desc));
-  req.bmRequestType = USBmakebmRequestType ((UInt8)kUSBIn,
-    (UInt8)kUSBStandard, (UInt8)kUSBDevice);
+  req.bmRequestType = USBmakebmRequestType (kUSBIn, kUSBStandard, kUSBDevice);
   req.bRequest      = kUSBRqGetDescriptor;
   req.wValue        = (UInt16)((LIBUSB_DT_STRING << 8) | string_index);
   req.wIndex        = langid;
@@ -1371,8 +1369,7 @@ static IOReturn darwin_request_descriptor (usb_device_t device, UInt8 desc, UInt
   memset (buffer, 0, buffer_size);
 
   /* Set up request for descriptor/ */
-  req.bmRequestType = USBmakebmRequestType((UInt8)kUSBIn,
-    (UInt8)kUSBStandard, (UInt8)kUSBDevice);
+  req.bmRequestType = USBmakebmRequestType(kUSBIn, kUSBStandard, kUSBDevice);
   req.bRequest      = kUSBRqGetDescriptor;
   req.wValue        = (UInt16)(desc << 8);
   req.wIndex        = desc_index;
